back/backend/requirements.txt
Django
djangorestframework
chromadb
sentence-transformers
requests
fuzzywuzzy
python-Levenshtein